+++ src/main/java/tavebalak/OTTify/user/Service/UserServiceImpl.java
@@ -6,11 +6,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import tavebalak.OTTify.community.dto.MyDiscussionDto;
+import tavebalak.OTTify.community.dto.response.MyDiscussionDto;
 import tavebalak.OTTify.community.entity.Community;
 import tavebalak.OTTify.community.repository.CommunityRepository;
 import tavebalak.OTTify.community.repository.ReplyRepository;
-import tavebalak.OTTify.review.dto.MyReviewDto;
+import tavebalak.OTTify.review.dto.response.MyReviewDto;
 import tavebalak.OTTify.review.entity.Review;
 import tavebalak.OTTify.review.entity.ReviewTag;
 import tavebalak.OTTify.review.repository.ReviewRepository;
